fix repeats, os paths, change command line args
multiple changes here. i've put back and re-looped in the number of
repeats, changed the chlorophyll decay time, removed the different named
rate schemes and replaced them with a required name arg which is the
name of the output folder, and added/removed various other command line
args. compiles and runs fine now.